+++ examples/iris_clustering_app.py
@@ -1,72 +1,3 @@
-# dash-template-rendering
-
-> Develop Dash layout with ease, using html template files.
-
-![PyPI - Licence](https://img.shields.io/pypi/l/dash-template-rendering)
-[![Current version on PyPI](https://img.shields.io/pypi/v/dash-template-rendering)](https://pypi.org/project/dash-template-rendering/)
-[![PyPI - Python Version](https://img.shields.io/pypi/pyversions/dash-template-rendering.svg?color=dark-green)](https://pypi.org/project/dash-template-rendering/)
-[![Build Status](https://github.com/pschleiter/dash-template-rendering/workflows/build/badge.svg)](https://github.com/pschleiter/dash-template-rendering/actions)
-[![codecov](https://codecov.io/gh/pschleiter/dash-template-rendering/branch/main/graph/badge.svg?token=148H4RN9NN)](https://codecov.io/gh/pschleiter/dash-template-rendering)
-
-dash_template_rendering is a simple [Dash](https://dash.plotly.com/) Python extension for rendering html content
-for the dashboard with [Jinja2](https://palletsprojects.com/p/jinja/). The rendered content can be used for the `dash.layout`
-and the return of callbacks.
-
-## Installation
-
-```
-$ pip install dash-template-rendering
-```
-
-## Examples
-
-### Minimal Dash App
-
-The [minimal Dash App](https://dash.plotly.com/minimal-app) implemented with the dash-template-rendering package.
-
-```python
-from dash import Dash, dcc, callback, Output, Input
-from dash_template_rendering import TemplateRenderer, render_dash_template_string
-import plotly.express as px
-import pandas as pd
-
-df = pd.read_csv(
-    "https://raw.githubusercontent.com/plotly/datasets/master/gapminder_unfiltered.csv"
-)
-
-app = Dash(__name__)
-TemplateRenderer(dash=app)
-
-app.layout = render_dash_template_string(
-    """
-<div>
-    <h1 style="text-align:center;">Title of Dash App</h1>
-    {{ dropdown | plotly }}
-    {{ graph | plotly }}
-</div>
-""",
-    dropdown=dcc.Dropdown(df.country.unique(), "Canada", id="dropdown-selection"),
-    graph=dcc.Graph(id="graph-content"),
-)
-
-
-@callback(Output("graph-content", "figure"), Input("dropdown-selection", "value"))
-def update_graph(value):
-    dff = df[df.country == value]
-    return px.line(dff, x="year", y="pop")
-
-
-if __name__ == "__main__":
-    app.run_server(debug=True)
-```
-
-### Advanced Example
-
-The full potential of this extension is applied, when combined with an CSS framework (e.g. [Bootstrap](https://getbootstrap.com/)).
-
-The [Iris k-means clustering Dash App](http://dash-bootstrap-components.opensource.faculty.ai/examples/iris/) implemented with the dash-template-rendering package.
-
-```python
 import dash
 import dash_bootstrap_components as dbc
 import pandas as pd
@@ -188,10 +119,3 @@ app.callback(Output("y-variable", "options"), [Input("x-variable", "value")])(
 
 if __name__ == "__main__":
     app.run_server(debug=True, port=8888)
-
-```
-
-## License
-
-This project is licensed under the MIT License (see the `LICENSE` file for
-details).
